second commit(fixes, upgrades)

+++ src/main/java/com/daniil/bank/demo/dal/repository/GuarantorRepository.java
@@ -1,12 +1,11 @@
 package com.daniil.bank.demo.dal.repository;
 
 import com.daniil.bank.demo.dal.entity.natural.Guarantor;
-import com.daniil.bank.demo.dal.entity.natural.Individual;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.validation.constraints.Size;
 
-public interface GuarantorRepository extends JpaRepository<Guarantor, Integer> {
+public interface GuarantorRepository extends JpaRepository<Guarantor, Long> {
 
     Guarantor findByPassportIDAndPassportSeries(@Size(min = 7, max = 7) String passportID, @Size(min = 2, max = 2) String passportSeries);
